modifico el int del numero 1 y numero 2 por float

+++ Calculadora_DavidPrado/src/main/java/calculadora_davidprado/calculadora_davidprado/Calculadora_DavidPrado.java
@@ -13,7 +13,7 @@ public class Calculadora_DavidPrado {
 
     static Scanner scanner = new Scanner(System.in);
     static int opcion = -1; //opciÃ³n del menÃº
-    static int numero1 = 0, numero2 = 0; //Variables de entrada
+    static float numero1 = 0, numero2 = 0; //Variables de entrada
     public static void main(String[] args) {
        
         System.out.println(" Hola Bienvenido usuario ");
